+++ src/components/TheControls/FormAddItem/index.js
@@ -1,0 +1,1 @@
+export { default } from "@/components/TheControls/FormAddItem/FormAddItem.vue";
